test(query): split query_parse_tests.rs at its section markers
crates/query/tests/query_parse_tests.rs was 691 lines, over the 500-line
convention for test files. Split it at its two existing section comments
into query_parse_tests.rs (parsing, 232 lines),
query_parse_error_tests.rs (error paths and fragments, 275 lines) and
query_parse_explain_tests.rs (explain directives and top-level
aggregates, 186 lines). Test functions are moved verbatim; only the two
new file headers and their `use query::parse_query;` lines are added.
49 tests before, 49 after (17 + 9 + 23).
